Update C++ implementation
